requirements.txt
matplotlib==3.5.2
numpy==1.21.5
ray==2.2.0
opencv_python==4.5.3.56
Pillow==9.2.0
scikit_learn==1.1.2
torch==1.12.1
torchvision==0.13.1
torch-summary==1.4.5
pandas==1.5.2
tabulate==0.9.0
tensorboardX==2.5.1
tensorboard==2.11.2
protobuf==3.20.0
open-clip-torch==2.11.1
